+++ src/main/java/GraduationProject/freelancermarket/service/concretes/EmployerManager.java
@@ -5,11 +5,14 @@ import java.util.List;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import GraduationProject.freelancermarket.core.business.BusinessRules;
 import GraduationProject.freelancermarket.entities.Employer;
 import GraduationProject.freelancermarket.model.dto.EmployerUpdateDto;
 import GraduationProject.freelancermarket.repository.EmployerRepository;
 import GraduationProject.freelancermarket.service.abstracts.EmployerService;
+import GraduationProject.freelancermarket.service.abstracts.TokenUserNameAndIdValidationService;
 import GraduationProject.freelancermarket.utils.DataResult;
+import GraduationProject.freelancermarket.utils.ErrorResult;
 import GraduationProject.freelancermarket.utils.Result;
 import GraduationProject.freelancermarket.utils.SuccessDataResult;
 import GraduationProject.freelancermarket.utils.SuccessResult;
@@ -20,6 +23,7 @@ import lombok.RequiredArgsConstructor;
 public class EmployerManager implements EmployerService {
 
 	private final EmployerRepository employerRepository;
+	private final TokenUserNameAndIdValidationService tokenUserNameAndIdValidationService;
 	private final ModelMapper modelMapper;
 
 	@Override
@@ -30,12 +34,20 @@ public class EmployerManager implements EmployerService {
 
 	@Override
 	public Result delete(int id) {
+		var businessRules = BusinessRules.run(userIdAndTokenUserNameVerification(id));
+		if (businessRules != null) {
+			return new ErrorResult(businessRules.getMessage());
+		}
 		employerRepository.deleteById(id);
 		return new SuccessResult("İşveren silindi");
 	}
 
 	@Override
 	public Result update(EmployerUpdateDto employerUpdateDto) {
+		var businessRules = BusinessRules.run(userIdAndTokenUserNameVerification(employerUpdateDto.getId()));
+		if (businessRules != null) {
+			return new ErrorResult(businessRules.getMessage());
+		}
 		Employer employer = modelMapper.map(employerUpdateDto, Employer.class);
 		employerRepository.save(employer);
 		return new SuccessResult("İşveren güncellendi");
@@ -49,6 +61,14 @@ public class EmployerManager implements EmployerService {
 	@Override
 	public DataResult<List<Employer>> getAll() {
 		return new SuccessDataResult<List<Employer>>(employerRepository.findAll(), "İşverenler listelendi");
+	}
+
+	public Result userIdAndTokenUserNameVerification(int userId) {
+		var result = tokenUserNameAndIdValidationService.userIdAndTokenUserNameVerification(userId);
+		if (!result.isSuccess()) {
+			return new ErrorResult(result.getMessage());
+		}
+		return new SuccessResult();
 	}
 
 }
