+++ src/main/java/GraduationProject/freelancermarket/service/concretes/SkillManager.java
@@ -5,11 +5,14 @@ import java.util.List;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import GraduationProject.freelancermarket.core.business.BusinessRules;
 import GraduationProject.freelancermarket.entities.Skill;
 import GraduationProject.freelancermarket.model.dto.SkillAddDto;
 import GraduationProject.freelancermarket.repository.SkillRepository;
 import GraduationProject.freelancermarket.service.abstracts.SkillService;
+import GraduationProject.freelancermarket.service.abstracts.TokenUserNameAndIdValidationService;
 import GraduationProject.freelancermarket.utils.DataResult;
+import GraduationProject.freelancermarket.utils.ErrorResult;
 import GraduationProject.freelancermarket.utils.Result;
 import GraduationProject.freelancermarket.utils.SuccessDataResult;
 import GraduationProject.freelancermarket.utils.SuccessResult;
@@ -20,10 +23,15 @@ import lombok.RequiredArgsConstructor;
 public class SkillManager implements SkillService {
 
 	private final SkillRepository skillRepository;
+	private final TokenUserNameAndIdValidationService tokenUserNameAndIdValidationService;
 	private final ModelMapper modelMapper;
 
 	@Override
 	public Result add(SkillAddDto skillAddDto) {
+		var businessRules = BusinessRules.run(userIdAndTokenUserNameVerification(skillAddDto.getFreelancerId()));
+		if (businessRules != null) {
+			return new ErrorResult(businessRules.getMessage());
+		}
 		Skill skill = modelMapper.map(skillAddDto, Skill.class);
 		skillRepository.save(skill);
 		return new SuccessResult("Yetenek eklendi");
@@ -31,6 +39,14 @@ public class SkillManager implements SkillService {
 
 	@Override
 	public Result delete(int id) {
+		var skill = skillRepository.findById(id).orElse(null);
+		if (skill == null) {
+			return new ErrorResult("Yetenek bulunamadı");
+		}
+		var businessRules = BusinessRules.run(userIdAndTokenUserNameVerification(skill.getFreelancerId()));
+		if (businessRules != null) {
+			return new ErrorResult(businessRules.getMessage());
+		}
 		skillRepository.deleteById(id);
 		return new SuccessResult("Yetenek silindi");
 	}
@@ -39,6 +55,14 @@ public class SkillManager implements SkillService {
 	public DataResult<List<Skill>> getByFreelancerId(int freelancerId) {
 		return new SuccessDataResult<List<Skill>>(skillRepository.findByFreelancerId(freelancerId),
 				"Freelancerın yetenekleri listelendi");
+	}
+
+	public Result userIdAndTokenUserNameVerification(int userId) {
+		var result = tokenUserNameAndIdValidationService.userIdAndTokenUserNameVerification(userId);
+		if (!result.isSuccess()) {
+			return new ErrorResult(result.getMessage());
+		}
+		return new SuccessResult();
 	}
 
 }
